+++ python/boomer/common/cpp/tuples.h
@@ -13,7 +13,7 @@
  * value in an array.
  */
 struct IndexedFloat32 {
-    intp index;
+    uint32 index;
     float32 value;
 };
 
@@ -23,7 +23,7 @@ struct IndexedFloat32 {
  */
 struct IndexedFloat32Array {
     IndexedFloat32* data;
-    intp numElements;
+    uint32 numElements;
 };
 
 /**
@@ -31,7 +31,7 @@ struct IndexedFloat32Array {
  * value in an array.
  */
 struct IndexedFloat64 {
-    intp index;
+    uint32 index;
     float64 value;
 };
